assignment_1/requirements.txt
psycopg2-binary
python-dotenv
orjson